requirements.txt
Django==1.11.2	
Pillow==4.1.1	
captcha==0.2.3	
django-allauth==0.32.0	
django-auto-urls==0.1	
django-celery==3.2.1	
django-crispy-forms==1.6.1	
django-mptt==0.8.7	
django-simple-captcha==0.5.5	
django-simple-serializer==2.0.7	
future==0.16.0	
requests==2.18.3
django-ipware	
